+++ sglang_omni/platforms/cuda.py
@@ -33,6 +33,21 @@ def _is_h20_device() -> bool:
         return False
 
 
+def _is_fp8_cutlass_moe_supported() -> bool:
+    """Mirror SGLang 0.5.16's CUTLASS FP8 MoE assertions."""
+    from sglang.srt.layers.quantization.fp8_utils import cutlass_fp8_supported
+    from sglang.srt.utils import (
+        is_sm90_supported,
+        is_sm100_supported,
+        is_sm120_supported,
+    )
+
+    return bool(
+        cutlass_fp8_supported()
+        and (is_sm90_supported() or is_sm100_supported() or is_sm120_supported())
+    )
+
+
 class CUDAOmniPlatform(CudaDeviceMixin, OmniPlatform):
     def get_stage_process_env(
         self,
